requirements.txt
python-dotenv
pillow
groq

# voice related libraries
ffmpeg    # ffmpeg is a powerful, open-source command-line tool used to process, convert, and manipulate audio and video files. In your case, it's essential for pydub to export audio in formats like MP3
SpeechRecognition     # To capture audio from microphone
pydub       # To convert and export audio to MP3
PyAudio    # Required for microphone input

# pipwin        #port audio


gtts # google text to speech
elevenlabs  # text to speech


streamlit
logging